refactor(graphs): extract shared hooks and utilities
Extract sorting logic and severity color mapping to reusable utilities. Create custom hook for sortable data to reduce code duplication across components.

Changes:
- Add useSortableData custom hook for data sorting logic
- Extract getSeverityColorByRiskScore to severity-utils
- Refactor HorizontalBarChart to use useSortableData hook
- Update ScatterPlot to use severity-utils
- Enhance AlertPill component with configurable props
- Improve ChartTooltip component structure
