+++ ui/components/graphs/shared/ChartTooltip.tsx
@@ -65,7 +65,10 @@ export function ChartTooltip({
       {data.newFindings !== undefined && data.newFindings > 0 && (
         <div className="mt-1 flex items-center gap-2">
           <Bell size={14} style={{ color: CHART_COLORS.textSecondary }} />
-          <span className="text-xs" style={{ color: CHART_COLORS.textSecondary }}>
+          <span
+            className="text-xs"
+            style={{ color: CHART_COLORS.textSecondary }}
+          >
             {data.newFindings} New Findings
           </span>
         </div>
@@ -74,7 +77,10 @@ export function ChartTooltip({
       {data.new !== undefined && data.new > 0 && (
         <div className="mt-1 flex items-center gap-2">
           <Bell size={14} style={{ color: CHART_COLORS.textSecondary }} />
-          <span className="text-xs" style={{ color: CHART_COLORS.textSecondary }}>
+          <span
+            className="text-xs"
+            style={{ color: CHART_COLORS.textSecondary }}
+          >
             {data.new} New
           </span>
         </div>
@@ -83,14 +89,20 @@ export function ChartTooltip({
       {data.muted !== undefined && data.muted > 0 && (
         <div className="mt-1 flex items-center gap-2">
           <VolumeX size={14} style={{ color: CHART_COLORS.textSecondary }} />
-          <span className="text-xs" style={{ color: CHART_COLORS.textSecondary }}>
+          <span
+            className="text-xs"
+            style={{ color: CHART_COLORS.textSecondary }}
+          >
             {data.muted} Muted
           </span>
         </div>
       )}
 
       {data.change !== undefined && (
-        <p className="mt-1 text-xs" style={{ color: CHART_COLORS.textSecondary }}>
+        <p
+          className="mt-1 text-xs"
+          style={{ color: CHART_COLORS.textSecondary }}
+        >
           <span className="font-bold">
             {data.change > 0 ? "+" : ""}
             {data.change}%
